backend/requirements.txt
argcomplete==1.10.3
asgiref==3.8.1
beautifulsoup4==4.8.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
compressed_rtf==1.0.6
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.18
extract-msg==0.28.7
idna==3.7
IMAPClient==2.1.0
lxml==5.2.1
mongoengine==0.28.2
olefile==0.47
pdfminer.six==20191110
pillow==10.3.0
pycryptodome==3.20.0
pymongo==4.6.3
python-pptx==0.6.23
pytz==2024.1
requests==2.31.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
textract==1.6.5
tzlocal==5.2
urllib3==2.2.1
xlrd==1.2.0
XlsxWriter==3.2.0
